git_info/requirements.txt

Django==2.1
djangorestframework==3.9.2
jsonfield==2.0.2
requests==2.21.0



